[MinGW] Execute cc using pex_one of libiberty
In Windows, execvp doesn't behave like what it did in Linux.

What we really need is that the original process is replaced with
new process, but this cannot be done using WIN32 API.

In Windows, the original behavior of execvp causes program 'make'
wrongly consider the executed program (wrapper) finish, so the
next statement in Makefile is executed, but the eventually executed
program (compiler) is not done yet.

To solve this racing, use the pex_one of libiberty, which waits
for the executed program's finish.

Note:
1. pex_one's usage is from gcc/gcc-ar.cc
2. gcc/system.h is included to get declaration
(e.g. WIFSIGNALED, WTERMSIG, strsignal)